add user in tasks and Logger Class

+++ src/tasks/tasks.service.ts
@@ -6,6 +6,8 @@ import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
 import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'src/auth/user.entity';
+import { GetUser } from 'src/auth/get-user.decorator';
 
 @Injectable()
 export class TasksService {
@@ -16,12 +18,12 @@ export class TasksService {
 
 
 
-  getTasks(filterDto : GetTasksFilterDto) :Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto);
+  getTasks(filterDto : GetTasksFilterDto, user: User) :Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user);
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 
     if(!found) {
       throw new NotFoundException(`Not found task with id: ${id}.`);
@@ -30,19 +32,20 @@ export class TasksService {
     return found;
   }
 
-  async cretaeTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.cretaeTask(createTaskDto);
+  async cretaeTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.cretaeTask(createTaskDto, user);
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id)
+  async deleteTask(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
+    
     if (result.affected === 0){
       throw new NotFoundException(`Not found task with id: ${id}.`);
     }
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task>{
-      const task = await this.getTaskById(id);
+  async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task>{
+      const task = await this.getTaskById(id, user);
       task.status = status;
       await task.save();
       return task;
